Add TodaysAnalyticsPage
